requirements.txt
torch>=2.0.0
diffusers>=0.24.0
transformers>=4.36.0
opencv-python
Pillow
gradio
accelerate
xformers ; platform_system != 'Windows'
